Alternate implementation on environment.
