+++ clients/admin-ui/src/app/store.ts
@@ -2,12 +2,9 @@ import {
   AnyAction,
   combineReducers,
   configureStore,
-  isRejectedWithValue,
-  Middleware,
   StateFromReducersMapObject,
 } from "@reduxjs/toolkit";
 import { setupListeners } from "@reduxjs/toolkit/query/react";
-import { createStandaloneToast } from "fidesui";
 import {
   FLUSH,
   PAUSE,
@@ -20,13 +17,13 @@ import {
 } from "redux-persist";
 import createWebStorage from "redux-persist/lib/storage/createWebStorage";
 
+import { rtkQueryErrorLogger, testRtkQueryErrorLogger } from "~/app/middleware";
 import { STORAGE_ROOT_KEY } from "~/constants";
 import { authSlice } from "~/features/auth";
 import { baseApi } from "~/features/common/api.slice";
 import { featuresSlice } from "~/features/common/features";
 import { healthApi } from "~/features/common/health.slice";
 import { dirtyFormsSlice } from "~/features/common/hooks/dirty-forms.slice";
-import { DEFAULT_TOAST_PARAMS } from "~/features/common/toast";
 import { configWizardSlice } from "~/features/config-wizard/config-wizard.slice";
 import { connectionTypeSlice } from "~/features/connection-type";
 import { tcfConfigSlice } from "~/features/consent-settings/tcf/tcf-config.slice";
@@ -48,10 +45,6 @@ import { dictSuggestionsSlice } from "~/features/system/dictionary-form/dict-sug
 import { taxonomySlice } from "~/features/taxonomy";
 import { datasetTestSlice } from "~/features/test-datasets";
 import { userManagementSlice } from "~/features/user-management";
-
-const { toast } = createStandaloneToast({
-  defaultOptions: DEFAULT_TOAST_PARAMS,
-});
 
 /**
  * To prevent the "redux-perist failed to create sync storage. falling back to noop storage"
@@ -110,7 +103,6 @@ const reducer = {
 export type RootState = StateFromReducersMapObject<typeof reducer>;
 
 const allReducers = combineReducers(reducer);
-
 const rootReducer = (state: RootState | undefined, action: AnyAction) => {
   let newState = state;
   if (action.type === "auth/logout") {
@@ -136,35 +128,14 @@ const persistConfig = {
   ],
 };
 
-export const persistedReducer = persistReducer(persistConfig, rootReducer);
-export const rtkQueryErrorLogger: Middleware = () => (next) => (action) => {
-  if (isRejectedWithValue(action)) {
-    const payload = action?.payload as any;
-    toast({
-      status: "error",
-      title: payload?.status ?? "An error occured",
-      description:
-        action?.error?.message ??
-        payload?.error ??
-        "An error occurred please check the console for more detail.",
-    });
-    // eslint-disable-next-line no-console
-    console.error(action.payload);
-  }
-  next(action);
-};
-
 const middleware = [baseApi.middleware, healthApi.middleware];
-
 if (process.env.NEXT_PUBLIC_APP_ENV !== "test") {
-  // Adding this middleware in `test` caused Cypress tests
-  // to take a lot longer than they otherwise would because errant
-  // notifications caused Cypress to be unable to click on elements.
-  // TODO: Conditionally enable this in tests so that we can start to
-  // clean up errors that are fired during tests.
   middleware.push(rtkQueryErrorLogger);
+} else {
+  middleware.push(testRtkQueryErrorLogger);
 }
 
+export const persistedReducer = persistReducer(persistConfig, rootReducer);
 export const makeStore = (
   preloadedState?: Parameters<typeof persistedReducer>[0],
 ) =>
